move helper func
